🔨 refac(user/table) Follow 위치 분리, 내부 구조 변경 필요 #80
