+++ src/main/java/Backend/HIFI/domain/follow/repository/FollowRepository.java
@@ -1,6 +1,6 @@
-package Backend.HIFI.domain.user.repository;
+package Backend.HIFI.domain.follow.repository;
 
-import Backend.HIFI.domain.user.entity.Follow;
+import Backend.HIFI.domain.follow.entity.Follow;
 import Backend.HIFI.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
